+++ src/app/shared/PersonSummary.svelte
@@ -15,7 +15,7 @@
   const showDetail = () => router.at("people").of(pubkey).open()
 </script>
 
-<div class="relative flex flex-grow flex-col gap-4 px-3 py-2">
+<div class="relative flex flex-grow flex-col gap-4">
   <div class="relative grid grid-cols-4 gap-4">
     <Anchor
       on:click={inert ? null : showDetail}
@@ -28,7 +28,7 @@
     </Anchor>
     {#if !hideActions}
       <div class="flex items-start justify-end">
-        <slot name="actions">
+        <slot name="actions" {pubkey}>
           <PersonActions {pubkey} />
         </slot>
       </div>