+++ src/app/shared/RelayCard.svelte
@@ -13,6 +13,7 @@
   export let showStatus = false
   export let hideActions = false
   export let showControls = false
+  export let inert = false
 </script>
 
 <div
@@ -20,12 +21,17 @@
   style={`border-left-color: ${hsl(stringToHue(relay.url))}`}>
   <div class="flex items-center justify-between gap-2">
     <div class="flex min-w-0 items-center gap-2 text-xl">
-      <i class={relay.url.startsWith("ws://") ? "fa fa-unlock" : "fa fa-lock"} />
-      <Anchor
-        href={router.at("relays").of(relay.url).toString()}
-        class="overflow-hidden text-ellipsis whitespace-nowrap">
-        {displayRelay(relay)}
-      </Anchor>
+      {#if inert}
+        <div class="overflow-hidden text-ellipsis whitespace-nowrap">
+          {displayRelay(relay)}
+        </div>
+      {:else}
+        <Anchor
+          href={router.at("relays").of(relay.url).toString()}
+          class="overflow-hidden text-ellipsis whitespace-nowrap">
+          {displayRelay(relay)}
+        </Anchor>
+      {/if}
       {#if showStatus && !getSetting("multiplextr_url")}
         <RelayStatus {relay} />
       {/if}