+++ src/engine/lists/commands.ts
@@ -2,9 +2,16 @@ import {pubkey} from "src/engine/session/state"
 import {createAndPublish, getClientTags} from "src/engine/network/utils"
 import {publishDeletion} from "src/engine/notes/commands"
 
-export const publishBookmarksList = (id, name, tags) => {
+export const publishBookmarksList = (id, title, description, tags) => {
   createAndPublish(30003, {
-    tags: [["d", id], ["name", name], ...getClientTags(), ...tags],
+    tags: [
+      ["d", id],
+      ["name", title],
+      ["title", title],
+      ["description", description],
+      ...getClientTags(),
+      ...tags,
+    ],
   })
 
   // migrate away from kind 30001